% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dnase}
\alias{dnase}
\title{Subset of \code{rhDNase} from the \code{survival} package}
\format{
A data frame with 767 observations on the following variables:
\describe{
\item{trt}{treatment arm: 0=placebo, 1= rhDNase}
\item{fev}{forced expriatory volume at enrollment, a measure of lung
capacity}
\item{infect}{an infection that required the use of intravenous
antibiotics}
\item{time}{difference between the date of entry into the study and
the date of last follow-up capped at 169 days}
}
}
\source{
\code{survival} package.
\url{https://cran.r-project.org/package=survival}
}
\usage{
dnase
}
\description{
The \code{dnase} dataset is a subset of the \code{rhDNase} dataset from the
\code{survival} package.
It is included in this package under the LGPL (\eqn{\geq2}) license.
}
\keyword{datasets}
