% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELBO.R
\name{elbo_cluster}
\alias{elbo_cluster}
\title{Calculates the variational Bayes convergence criteria, evidence lower
bound (ELBO), optimized in \code{survregVB.frailty.fit}.}
\usage{
elbo_cluster(
  y,
  X,
  delta,
  alpha_0,
  omega_0,
  mu_0,
  v_0,
  lambda_0,
  eta_0,
  alpha,
  omega,
  mu,
  Sigma,
  tau,
  sigma,
  lambda,
  eta,
  expectation_b,
  cluster
)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A matrix of predictors (covariates), including an intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{alpha_0}{Hyperparameter \eqn{\alpha_0} of the prior
distribution of \emph{b}.}

\item{omega_0}{Hyperparameter \eqn{\omega_0} of the prior
distribution of \emph{b}.}

\item{mu_0}{Hyperparameter \eqn{\mu_0} of the prior distribution of
\emph{β}.}

\item{v_0}{Precision hyperparameter \eqn{v_0} of the prior
distribution of \emph{β}.}

\item{lambda_0}{Hyperparameter \eqn{\lambda_0} of the prior distribution
of \eqn{sigma^2_{\gamma}}.}

\item{eta_0}{Hyperparameter \eqn{\eta_0} of the prior distribution of
\eqn{sigma^2_{\gamma}}.}

\item{alpha}{Parameter \eqn{\alpha^*} of \eqn{q^*(b)}.}

\item{omega}{Parameter \eqn{\omega^*} of \eqn{q^*(b)}.}

\item{mu}{The current value of the parameter \eqn{\mu^*} of \eqn{q^*(\beta)}.}

\item{Sigma}{Parameter \eqn{Sigma^*} of \eqn{q^*(\beta)}.}

\item{tau}{Parameter \eqn{\tau^*} of \eqn{q^*(\gamma)} for all clusters.}

\item{sigma}{Parameter \eqn{\sigma^{2*}} of \eqn{q^*(\gamma)} for all
clusters.}

\item{lambda}{Parameter \eqn{\lambda^*} of \eqn{q^*(sigma^2_{\gamma})}
for all clusters.}

\item{eta}{Parameter \eqn{\eta^*} of \eqn{q^*(sigma^2_{\gamma})} for
all clusters.}

\item{expectation_b}{The expected value of b.}

\item{cluster}{A numeric vector indicating the cluster assignment for
each observation.}
}
\value{
The evidence lower bound (ELBO).
}
\description{
Calculates the variational Bayes convergence criteria, evidence lower
bound (ELBO), optimized in \code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.fit}}
}
