% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{eta_star}
\alias{eta_star}
\title{Calculates parameter \eqn{eta^*} of \eqn{q^*(sigma^2_{\gamma})} for all
clusters to optimize the evidence based lower bound (ELBO) in
\code{survregVB.frailty.fit}.}
\usage{
eta_star(eta_0, tau, sigma)
}
\arguments{
\item{eta_0}{Hyperparameter \eqn{\eta_0} of the prior distribution of
\eqn{sigma^2_{\gamma}}.}

\item{tau}{Parameter \eqn{\tau^*} of \eqn{q^*(\gamma)} for all clusters.}

\item{sigma}{Parameter \eqn{\sigma^{2*}} of \eqn{q^*(\gamma)} for all
clusters.}
}
\value{
Parameter \eqn{eta^*} of \eqn{q^*(sigma^2_{\gamma})} for all
clusters.
}
\description{
Calculates parameter \eqn{eta^*} of \eqn{q^*(sigma^2_{\gamma})} for all
clusters to optimize the evidence based lower bound (ELBO) in
\code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.frailty.fit}}
}
