% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{lambda_star}
\alias{lambda_star}
\title{Calculates parameter \eqn{\lambda^*} of \eqn{q^*(\sigma^2_{\gamma})} to
optimize the evidence based lower bound (ELBO) in
\code{survregVB.frailty.fit}.}
\usage{
lambda_star(lambda_0, K)
}
\arguments{
\item{lambda_0}{The shape hyperparameter \eqn{\lambda_0} of the prior
distribution of the frailty variance, \eqn{\sigma_\gamma^2}.}

\item{K}{The number of clusters.}
}
\value{
Parameter \eqn{\lambda^*} of \eqn{q^*(\sigma^2_{\gamma})}.
}
\description{
Calculates parameter \eqn{\lambda^*} of \eqn{q^*(\sigma^2_{\gamma})} to
optimize the evidence based lower bound (ELBO) in
\code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.frailty.fit}}
}
