% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_cancer}
\alias{lung_cancer}
\title{Subset of GSE102287: African American (AA) Patients}
\format{
A data frame with 60 observations on selected patient
characteristics:
\describe{
\item{patient}{Patient identification number.}
\item{age}{Patient age.}
\item{Stage}{Lung cancer stage (I, II, III).}
\item{time}{Survival time in days.}
\item{gender}{Gender of the patient.}
\item{smoking}{0 = Never smoked, 1 = Has smoked.}
\item{status}{0 = Alive, 1 = Death due to lung cancer.}
}
}
\source{
Gene Expression Omnibus (GEO), Accession: GSE102287.
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE102287}
}
\usage{
lung_cancer
}
\description{
This dataset is a subset of the GSE102287 dataset that includes only
characteristics of patients who are identified as African American (AA).
}
\references{
Mitchell, K. A., Zingone, A., Toulabi, L., Boeckelman, J.,
& Ryan, B. M. (2017). Comparative Transcriptome Profiling Reveals Coding
and Noncoding RNA Differences in NSCLC from African Americans and European
Americans. Clinical cancer research: an official journal of the American
Association for Cancer Research, 23(23), 7412–7425.
doi:10.1158/1078-0432.CCR-17-0527.
}
\keyword{datasets}
