% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{Sigma_star_cluster}
\alias{Sigma_star_cluster}
\title{Calculates parameter \eqn{\Sigma^*} of \eqn{q^*(\beta)} to optimize the
evidence based lower bound (ELBO) in \code{survregVB.frailty.fit}.}
\usage{
Sigma_star_cluster(
  y,
  X,
  delta,
  v_0,
  alpha,
  omega,
  mu,
  tau,
  expectation_b,
  cluster
)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A design matrix including covariates with first column of ones
to represent the intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{v_0}{The precision (inverse variance) hyperparameter \eqn{v_0},
of the prior distribution of the vector of coefficients, \eqn{\beta}.}

\item{alpha}{The shape parameter \eqn{\alpha^*} of \eqn{q^*(b)}.}

\item{omega}{The scale parameter \eqn{\omega^*} of \eqn{q^*(b)}.}

\item{mu}{Parameter \eqn{\mu^*} of \eqn{q^*(\beta)}, a vector of means.}

\item{tau}{Parameter \eqn{\tau^*} of \eqn{q^*(\gamma_i)}, a vector of
means.}

\item{expectation_b}{The expected value of \emph{b}.}

\item{cluster}{A numeric vector indicating the cluster assignment for
each observation.}
}
\value{
Parameter \eqn{\Sigma^*} of \eqn{q^*(\beta)}.
}
\description{
Calculates parameter \eqn{\Sigma^*} of \eqn{q^*(\beta)} to optimize the
evidence based lower bound (ELBO) in \code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.frailty.fit}}
}
