% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survregVB.R
\name{summary_table}
\alias{summary_table}
\title{Produces a summary table containing the value and confidence intervals
for an Inverse-Gamma posterior distribution.}
\usage{
summary_table(shape, scale)
}
\arguments{
\item{shape}{The shape parameter of the distribution.}

\item{scale}{The scale parameter of the distribution.}
}
\description{
Produces a summary table containing the value and confidence intervals
for an Inverse-Gamma posterior distribution.
}
