% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_elbo.R
\name{elbo}
\alias{elbo}
\title{Calculates the variational Bayes convergence criteria, evidence lower
bound (ELBO), optimized in \code{survregVB.fit}.}
\usage{
elbo(
  y,
  X,
  delta,
  alpha_0,
  omega_0,
  mu_0,
  v_0,
  alpha,
  omega,
  curr_mu,
  Sigma,
  expectation_b
)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A matrix of predictors (covariates), including an intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{alpha_0}{Hyperparameter \eqn{\alpha_0} of the prior
distribution of \emph{b}.}

\item{omega_0}{Hyperparameter \eqn{\omega_0} of the prior
distribution of \emph{b}.}

\item{mu_0}{Hyperparameter \eqn{\mu_0} of the prior distribution of
\emph{β}.}

\item{v_0}{Precision hyperparameter \eqn{v_0} of the prior
distribution of \emph{β}.}

\item{alpha}{Parameter \eqn{\alpha^*} of the approximate posterior
distribution of \emph{b}.}

\item{omega}{Parameter \eqn{\omega^*} of the approximate posterior
distribution of \emph{b}.}

\item{curr_mu}{The current value of the parameter \eqn{\mu^*} of the
approximate posterior distribution of \emph{β}.}

\item{Sigma}{Parameter \eqn{Sigma^*} of the approximate posterior
distribution of\emph{β}.}

\item{expectation_b}{The expected value of b.}
}
\value{
The evidence lower bound (ELBO).
}
\description{
Calculates the variational Bayes convergence criteria, evidence lower
bound (ELBO), optimized in \code{survregVB.fit}.
}
\seealso{
\code{\link{survregVB.fit}}
}
