% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{mu_star}
\alias{mu_star}
\title{Calculates parameter \eqn{\mu^*} of the approximate posterior
distribution of \emph{β} in order to optimize the evidence based
lower bound (ELBO) in \code{survregVB.fit}.}
\usage{
mu_star(y, X, delta, mu_0, v_0, alpha, omega, curr_mu, Sigma, expectation_b)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A matrix of predictors (covariates), including an intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{mu_0}{Hyperparameter \eqn{\mu_0} of the prior distribution of
\emph{β}.}

\item{v_0}{Hyperparameter \eqn{v_0} of the prior distribution of
\emph{β}.}

\item{alpha}{Parameter \eqn{\alpha^*} of the approximate posterior
distribution of \emph{b}.}

\item{omega}{Parameter \eqn{\omega^*} of the approximate posterior
distribution of \emph{b}.}

\item{curr_mu}{The current value of the parameter \eqn{\mu^*} of the
approximate posterior distribution of \emph{β}.}

\item{Sigma}{Parameter \eqn{\Sigma^*} of the approximate posterior
distribution of\emph{β}.}

\item{expectation_b}{The expected value of b.}
}
\value{
Parameter \eqn{\mu^*} of the approximate posterior
distribution of \emph{β}.
}
\description{
Calculates parameter \eqn{\mu^*} of the approximate posterior
distribution of \emph{β} in order to optimize the evidence based
lower bound (ELBO) in \code{survregVB.fit}.
}
\seealso{
\code{\link{survregVB.fit}}

\code{\link{elbo}}
}
