% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{omega_star_cluster}
\alias{omega_star_cluster}
\title{Calculates parameter \eqn{\omega^*} of the approximate posterior
distribution of \emph{b} in order to optimize the evidence based
lower bound (ELBO) in \code{survregVB.frailty.fit}.}
\usage{
omega_star_cluster(y, X, delta, omega_0, mu, tau, expectation_b, cluster_index)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A matrix of predictors (covariates), including an intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{omega_0}{Hyperparameter \eqn{\omega_0} of the prior
distribution of \emph{b}.}

\item{mu}{Parameter \eqn{\mu^*} of the approximate posterior
distribution of \emph{β}.}

\item{expectation_b}{The expected value of b.}
}
\value{
Parameter \eqn{\omega^*} of the approximate posterior
distribution of \emph{b}.
}
\description{
Calculates parameter \eqn{\omega^*} of the approximate posterior
distribution of \emph{b} in order to optimize the evidence based
lower bound (ELBO) in \code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.fit}}
}
