% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{sigma_squared_star}
\alias{sigma_squared_star}
\title{Calculates parameter \eqn{sigma^{2*}} of \eqn{q^*(\gamma)} for all clusters
to optimize the evidence based lower bound (ELBO) in \code{survregVB.frailty.fit}.}
\usage{
sigma_squared_star(
  y,
  X,
  delta,
  alpha,
  omega,
  mu,
  tau,
  lambda,
  eta,
  expectation_b,
  cluster
)
}
\arguments{
\item{y}{A vector of observed log-transformed survival times.}

\item{X}{A matrix of predictors (covariates), including an intercept.}

\item{delta}{A binary vector indicating right censoring.}

\item{alpha}{Parameter \eqn{\alpha^*} of \eqn{q^*(b)}.}

\item{omega}{Parameter \eqn{\omega^*} of \eqn{q^*(b)}.}

\item{mu}{The parameter \eqn{\mu^*} of \eqn{q^*(\beta)}.}

\item{tau}{Parameter \eqn{\tau^*} of \eqn{q^*(\gamma)} for all clusters.}

\item{lambda}{Parameter \eqn{\lambda^*} of \eqn{q^*(sigma^2_{\gamma})}
for all clusters.}

\item{eta}{Parameter \eqn{\eta^*} of \eqn{q^*(sigma^2_{\gamma})} for
all clusters.}

\item{expectation_b}{The expected value of b.}

\item{cluster}{A numeric vector indicating the cluster assignment for
each observation.}
}
\value{
Parameter \eqn{\sigma^{2*}} of \eqn{q^*(\gamma)} for all clusters.
}
\description{
Calculates parameter \eqn{sigma^{2*}} of \eqn{q^*(\gamma)} for all clusters
to optimize the evidence based lower bound (ELBO) in \code{survregVB.frailty.fit}.
}
\seealso{
\code{\link{survregVB.frailty.fit}}
}
