% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation_nofrailty}
\alias{simulation_nofrailty}
\title{Simulated data without shared frailty effects to model unclustered
time-to-event data.}
\format{
A dataframe with 300 observations.
\describe{
\item{x1}{Continuous covariate from \code{N(1, 0.2^2)}}
\item{x2}{Binary covariate from \code{Bernoulli(0.5)}}
\item{Time}{True survival time}
\item{Time.10}{Observed survival time accounting for uniformly distributed
right censoring time from \code{uniform(0,48)}}
\item{Time.30}{Observed survival time accounting for uniformly distributed
right censoring time from \code{uniform(0,17)}}
\item{delta}{Event indicator for uncensored data (always 1 in this
simulation.)}
\item{delta.10}{Event indicator for T.10 (1 = event, 0 =
censored).}
\item{delta.30}{Event indicator for T.30 (1 = event, 0 =
censored).}
}
@references Xian, C., Souza, C. P. E. de, He, W., Rodrigues, F. F.,
& Tian, R. (2024). Variational Bayesian analysis of survival data
using a log-logistic accelerated failure time model. Statistics and
Computing, 34(2). https://doi.org/10.1007/s11222-023-10365-6
}
\usage{
simulation_nofrailty
}
\description{
Simulated data without shared frailty effects to model unclustered
time-to-event data.
}
\keyword{datasets}
