% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.survregVB.R
\name{summary.survregVB}
\alias{summary.survregVB}
\alias{print.summary.survregVB}
\title{Summary for Variational Bayes log-logistic AFT models.}
\usage{
\method{summary}{survregVB}(object, ci = 0.95, ...)
}
\arguments{
\item{object}{The result of a \code{survregVB} fit.}

\item{ci}{The significance level for the credible intervals.
(Default:0.95).}

\item{\dots}{For future arguments.}
}
\value{
An object of class \code{summary.survregVB} with components:
\itemize{
\item \code{call}: The function call used to invoke the
\code{survregVB} method.
\item \code{ELBO}: The final value of the Evidence Lower Bound
(ELBO) at the last iteration.
\item \code{alpha}: The updated parameter \eqn{\alpha} of the
approximate posterior distribution of the scale parameter.
\item \code{omega}: The updated parameter \eqn{\omega} of the
approximate posterior distribution of the scale parameter.
\item \code{mu}: The updated vector of means \eqn{\mu} of the
approximate posterior distribution of the regression coefficients.
\item \code{Sigma}: The updated covariance matrix \eqn{\Sigma} of
the approximate posterior distribution of the regression
coefficients.
\item \code{tau}: The updated parameter \eqn{\tau} of the approximate
posterior distribution of the random intercept.
\item \code{sigma}: The updated parameter \eqn{\sigma^2} of the approximate
posterior distribution of the random intercept.
\item \code{lambda}: The updated parameter \eqn{\lambda} of the approximate
posterior distribution of the variance of the random intercept.
\item \code{eta}: The updated parameter \eqn{\eta} of the approximate
posterior distribution of the variance of the random intercept.
\item \code{na.action}: A missing-data filter function, applied to
the model.frame, after any subset argument has been used.
\item \code{iterations}: The number of iterations performed by the
VB algorithm: before converging or reaching \code{max_iteration}.
\item \code{n}: The number of observations.
\item \code{coefficients}: A matrix with one row for each
coefficient, and columns containing:
\itemize{
\item \code{Value}: The estimated value of the coefficient.
\item \code{Lower CI}: The lower bound of the credible interval
associated with the coefficient.
\item \code{Upper CI}: The upper bound of the credible interval
associated with the coefficient.
}
\item \code{scale}: A vector containing:
\itemize{
\item \code{Value}: The estimated value of the scale parameter.
\item \code{Lower CI}: The lower bound of the credible interval
associated with the scale parameter.
\item \code{Upper CI}: The upper bound of the credible interval
associated with the scale parameter.
}
}
}
\description{
Produces a summary of a fitted Variational Bayes Parametric Survival
Regression Model for a Log-Logistic AFT Model
}
\seealso{
\code{\link{survregVB}}
}
