% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survregVB.fit.R
\name{survregVB.object}
\alias{survregVB.object}
\alias{print.survregVB}
\title{Variational Bayesian Analysis of Survival Data Using a Log-Logistic
Accelerated Failure Time Model}
\description{
Called by \code{survregVB} to do the actual parameter and ELBO
computations.
}
\details{
Objects of this class have methods for the functions \code{print} and
\code{summary}. The components of this class are:
\itemize{
\item \code{ELBO}: The final value of the Evidence Lower Bound
(ELBO) at the last iteration.
\item \code{alpha}: The updated parameter \eqn{\alpha} of the
approximate posterior distribution of \emph{b}.
\item \code{omega}: The updated parameter \eqn{\omega} of the
approximate posterior distribution of \emph{b}.
\item \code{mu}: The updated vector of means \eqn{\mu} of the
approximate posterior distribution of \emph{β}.
\item \code{Sigma}: The updated covariance matrix \eqn{\Sigma} of
the approximate posterior distribution of \emph{β}.
\item \code{na.action}: A missing-data filter function, applied to
the model.frame, after any subset argument has been used.
\item \code{iterations}: The number of iterations performed by the
VB algorithm: before converging or reaching \code{max_iteration}.
\item \code{n}: The number of observations.
\item \code{call}: The function call used to invoke the
\code{survregVB} method.
\item \code{not_converged}: A boolean indicating if the algorithm
converged.
\itemize{
\item \code{TRUE}: If the algorithm did not converge prior to
achieving \code{max_iteration}.
\item \code{NULL}: If the algorithm converged successfully.
}
}
}
