% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survregVB.R
\name{survregVB.object}
\alias{survregVB.object}
\alias{print.survregVB}
\title{Variational Bayes Accelererated Failure Time Survival Model Object}
\description{
This class of objects is returned by the survregVB function to represent
a fitted parametric log-logistic accelerated failure time (AFT) survival
model. Objects of this class have methods for the functions \code{print}
and \code{summary}.
}
\details{
The components of this class are:
\itemize{
\item \code{ELBO}: The final value of the Evidence Lower Bound (ELBO)
at the last iteration.
\item \code{alpha}: The updated shape parameter \eqn{\alpha} of the
approximate posterior distribution of \emph{b}.
\item \code{omega}: The updated scale parameter \eqn{\omega} of the
approximate posterior distribution of \emph{b}.
\item \code{mu}: The updated vector of means \eqn{\mu} of the
approximate posterior distribution of \emph{β}.
\item \code{Sigma}: The updated covariance matrix \eqn{\Sigma} of the
approximate posterior distribution of \emph{β}.
\item \code{na.action}: A missing-data filter function, applied to the
\code{model.frame}, after any subset argument has been used.
\item \code{iterations}: The number of iterations performed by the VB
algorithm: before converging or reaching \code{max_iteration}.
\item \code{n}: The number of observations.
\item \code{call}: The function call used to invoke the \code{survregVB}
method.
\item \code{not_converged}: A boolean indicating if the algorithm converged.
\itemize{
\item \code{TRUE}: If the algorithm did not converge prior to
achieving \code{max_iteration}.
\item \code{NULL}: If the algorithm converged successfully.
}
}

The following components are present if \code{survregVB} was called with
the cluster argument (shared frailty):
\itemize{
\item \code{clustered}: A boolean indicating that the model was clustered.
\item \code{tau}: The updated vector of variance \eqn{\tau} of the
approximate posterior distribution of \eqn{\gamma_i} for \eqn{i=1,...,K}
clusters.
\item \code{sigma}: The updated vector of means \eqn{\sigma^2_i}
of the approximate posterior distribution of \eqn{\gamma_i} for
\eqn{i=1,...,K} clusters.
\item \code{lambda}: The updated shape parameter \eqn{\lambda_i} of
the approximate posterior distribution of \eqn{omega^2_{\gamma_i}}
for \eqn{i=1,...,K} clusters.
\item \code{eta}: The updated scale parameter \eqn{\eta_i} of the
approximated posterior distribution of \eqn{omega^2_{\gamma_i}} for
\eqn{i=1,...,K} clusters.
}
}
